.github/scripts/workflow_rerun/requirements.txt
PyGithub==2.8.1
requests==2.31.0
psycopg2-binary==2.9.9